Updated TreeNode.java and created Tree.java and TreeTest.java.
Updated TreeNode to work with String items instead of a generic data type. Implemented Tree.java: has methods to create a tree with a childless root, root with children, get the height of the tree, and construct a perfect tree of a given height and with given children. Created TreeTest.java to test usage of the tree.